ai_models/parth/ai_models/sarthi/model.py
import requests
import PyPDF2
import docx
import speech_recognition as sr

class SarthiAI:
    def __init__(self):
        pass

    def search_internet(self, query):
        # Placeholder for internet search functionality
        # You can use APIs like Google Custom Search JSON API
        response = requests.get(f"https://www.googleapis.com/customsearch/v1?q={query}&key=YOUR_API_KEY")
        return response.json()

    def read_pdf(self, file_path):
        # Function to read PDF files
        with open(file_path, 'rb') as file:
            reader = PyPDF2.PdfFileReader(file)
            text = ""
            for page_num in range(reader.numPages):
                page = reader.getPage(page_num)
                text += page.extract_text()
        return text

    def read_docx(self, file_path):
        # Function to read DOCX files
        doc = docx.Document(file_path)
        text = "\n".join([para.text for para in doc.paragraphs])
        return text

    def encode_audio(self, file_path):
        # Function to encode audio files
        recognizer = sr.Recognizer()
        with sr.AudioFile(file_path) as source:
            audio = recognizer.record(source)
            text = recognizer.recognize_google(audio)
        return text

# Example usage
if __name__ == "__main__":
    ai = SarthiAI()
    print(ai.search_internet("Python programming"))
    print(ai.read_pdf("example.pdf"))
    print(ai.read_docx("example.docx"))
    print(ai.encode_audio("example.wav"))
    
else:
    ai = SarthiAI()

