ai_models/parth/ai_models/sarthi/requirements.txt
SpeechRecognition
python-docx
PyPDF2
pydub
torch
transformers
pyaudio